Ignore EINTR signal for select systemcall.
Here select is a blocking call since no timeout is specified,
so we have chances of interrupting via EINTR.
When EINTR is recived, switchlink_main thread is exiting.
This causes infrap4d to not process any netlink events and lnp.p4
usecases will fail.

With the fix, we are ignoring EINTR error for Select systemcall

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>