Program nexthop table as a ternary match. (#59)
The P4 program has been modified to move the NextHop table from
the SEM block (exact match) to the WCM block (ternary match). This
requires a corresponding change in the Kernel Monitor.

This is a breaking change. Newer versions of krnlmon are incompatible
with older versions of the P4 program.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>
Co-authored-by: Derek G Foster <derek.foster@intel.com>