Add support for include-what-you-use (#61)
include-what-you-use is a utility developed at Google that analyzes C and C++ source files, looking for violations of the Include What You Use rule, and recommends fixes for them.

"The main goal of include-what-you-use is to remove superfluous #includes. It does this both by figuring out what #includes are not actually needed for this file (for both .cc and .h files), and replacing #includes with forward-declares when possible."

This CL implements a cmake option that can be used to run include-what-you-use on krnlmon. It requires standalone build support, plus a trivial change to FindDpdkDriver in order to analyze krnlmon for DPDK.

Signed-off-by: Derek G Foster <derek.foster@intel.com>