Implement P4 Driver Abstraction Layer (PDAL)
Recent changes to the ES2K P4 SDE made it incompatible with the
legacy SDE interface used by the DPDK and Tofino SDEs. PDAL
provides an abstract interface that allows krnlmon to communicate
transparently with either SDE.

The initial implementation is fairly simple. It provides a subset
of the features (just what's needed to support krnlmon), and it
leverages the fact that symbols have simply been renamed and are
directly equivalent. This is a translucent abstraction, not an
opaque one (just enough to get by).

Signed-off-by: Derek G Foster <derek.foster@intel.com>