Implement P4 SDE Abstraction Layer (switchsde)
Recent changes to the ES2K SDE made it incompatible with the
legacy interface used by the DPDK and Tofino SDEs. `switchsde`
provides an abstract interface that allows krnlmon to interact
transparently with both the legacy SDEs and the ES2K SDE.

The initial implementation is fairly simple. It provides only
the features that are needed to support krnlmon; and it takes
advantage of the fact that the symbols have simply been renamed,
so no data conversions are necessary (yet).

Signed-off-by: Derek G Foster <derek.foster@intel.com>