Krnlmon changes to support linux_networking-v3 (Combined Recipe) (#84)
* Linux networking V3 support

This PR inroduces support for linux networking version V3 which addresses
scalability issues in V2 version. Mainly tunnel and and forwarding tables
are moved to be exact match type to provide a scale of upto 1M entries. And
lag and ecmp tables are moved to WCM since scale of 1K is enough for these cases.

We have conditionalized V2 code under LNW_V2 and V3 code under LNW_V3.
The intention is to provide side-by-side support for V2 and V3. Also,
Bazel build will be updated for V3 codebase in a subsequent update. Once V3
version is stable, plan is to deprecate V2 version

Signed-off-by: nupuruttarwar <nupur.uttarwar@intel.com>