Populate corret RMAC handle when MAC address of bridge is modified
When an OvS bridge is created, it is assigned with a MAC address A.
But, when an IDPF interface is added to the OvS bridge, the MAC address
is
modified from A to B. This B MAC address is same as the IDPF MAC
address. P4CP should use this new B MAC address to derive port_id, which
eventually be used when underlay ARP & nexthop are programmed to fetch
physical_port_id.
Currently krnlmon handles MAC change, but doesnt populate port_id for
the new MAC B. Changes are made to fix it.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>