Fix error in standalone cmake build
The krnlmon_unit_tests workflow fails with the following message:

  CMake Error at CMakeLists.txt:69 (set_install_rpath):
    Unknown CMake command "set_install_rpath".

This is because set_install_rpath() is defined in the top-level
networking-recipe CMakeLists.txt file, which is not executed in
standalone build mode.

Addressed the issue by moving the function definition to its own
file (SetInstallPath.cmake) and including it from standalone.cmake
in the krnlmon build.

Note that this commit is dependent on a corresponding commit in
the networking-recipe repository.

Signed-off-by: Derek Foster <derek.foster@intel.com>